requirements.txt
# whisperflow dependencies
# install with: pip install -r requirements.txt

# OpenAI SDK — transcription API
openai>=1.30.0

# Global hotkey listener (works across all macOS apps)
pynput>=1.7.6

# Microphone recording
sounddevice>=0.4.6
soundfile>=0.12.1

# Audio buffer handling
numpy>=1.26.0

# Paste text into any app
pyperclip>=1.8.2
pyautogui>=0.9.54

# .env file support
python-dotenv>=1.0.0

# macOS menu bar + native notifications
rumps>=0.4.0
